requirements.txt
python-dotenv
uvicorn[standard]
fastapi[standard]
openai
black
requests
psycopg2-binary
py-AutoClean
pandas